msm: camera2: cpp: fix to release duplicate buffer
The CPP driver can support scheduling of two output frames
in duplication mode. In the event of error, in case first
output buffer is acquired, it needs to be released back to
free queue. If is it not released, the acquired buffer will
never be used for circulation and gets lost.

Change-Id: I7891586b6c8902b4a1301627f22c3c7ec429d025
Signed-off-by: Rajakumar Govindaram <rajakuma@codeaurora.org>