samsung: removed sim_slot count check
The sim-slot check driver auto detects the device's sim slot count
by checking the sim slot pin with gpio direction input.
However, this driver highly depends to gpio features that are not
available in arm64. Additionally the driver has a risk of potential
kernel panics if gpio fails to find the sim slot pin (for example
if the sim slot is broken or missing hardware-side).

This driver is not available in later samsung devices, therefore
it's safe to remove it. In order to use multiple sims the
configuration has to be set by user space. Example for dual sim:

[Android build.prop]
ro.multisim.simslotcount=2
persist.radio.multisim.config=dsds